更新BrandID

+++ BabySharing/Controller/Profile/PublishActivity/CarePublishPrepare/CareAgeChooseView/AYCareAgeChooseView.h
@@ -14,7 +14,7 @@
 @interface AYCareAgeChooseView : UIView <AYAgePickViewDelegate>
 
 @property(nonatomic, strong) AYAgePickView *pickView;
-@property(nonatomic, readonly) NSInteger ageMin;
-@property(nonatomic, readonly) NSInteger ageMax;
+@property(nonatomic, readonly) double ageMin;
+@property(nonatomic, readonly) double ageMax;
 
 @end
